Moved styles to files
