+++ components/Snippet/index.js
@@ -1,0 +1,8 @@
+import {styles} from './styles';
+
+export default ({children}) => (
+    <div>
+        {children}
+        <style jsx>{styles}</style>
+    </div>
+);
