+++ package.json
@@ -9,6 +9,7 @@
     "next": "latest",
     "react": "latest",
     "react-dom": "latest",
+    "styled-jsx": "^3.2.1",
     "url-parse": "^1.4.7"
   },
   "devDependencies": {
